视频模块-根据id获取用户的全部视频列表

+++ service/service-video/src/main/java/com/linyilinyi/video/service/VideoService.java
@@ -31,5 +31,5 @@ public interface VideoService extends IService<Video> {
 
     String updateVideo(Video video);
 
-    List<Video> getVideoListByUserId(Long userId);
+    List<Video> getVideoListByUserId(Integer userId);
 }
